+++ client/client.h
@@ -153,7 +153,7 @@ typedef struct
 	int			timeoutcount;
 
 	int			timedemo_frames;
-	int			timedemo_start;
+	double		timedemo_start;
 
 	qboolean	refresh_prepped;	// false if on new level or new ref dll
 	qboolean	sound_prepped;		// ambient sounds can start
